Allow passing a custom application ID
With this the user can have a custom title for the status.
Since the status is treated like a game, something like
"Playing Music", where the application name is "Music",
might be more desirable than "Playing TIDAL".
especially since "Listening to TIDAL" seems to
not be supported (#77) by the Game SDK.